rename merkle tree to dense merkle
