+++ stark/src/dense_merkle_tree/mod.rs
@@ -1,0 +1,2 @@
+mod merkle;
+mod util;
