+++ stark/src/dense_merkle_tree/util.rs
