+++ stark/src/sparse_merkle_tree/mod.rs
@@ -1,0 +1,1 @@
+mod merkle;
